% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_geo.R
\name{census_geo}
\alias{census_geo}
\title{Geocode Addresses Batches Using the Census Bureau API}
\usage{
census_geo(.data, sf = FALSE)
}
\arguments{
\item{.data}{A properly formatted data.frame containing addresses to geocode}

\item{sf}{If TRUE, returns data in an SF class object}
}
\description{
It is recommended that `census_geo()` be used in a pipe with `census_prep()`. See `vignette("censusxy")` for more details and example code.
When 10,000 or more addresses are specified, the function will parse the data into smaller chunks in order to overcome the limit of 10,000 requests. This functionality should be used responsibly.
}
