% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder.R
\name{census_geocoder}
\alias{census_geocoder}
\title{This is the internal function for geocoding
much of the code is borrowed from https://github.com/dmwelgus/MapChi}
\usage{
census_geocoder(df)
}
\description{
This is the internal function for geocoding
much of the code is borrowed from https://github.com/dmwelgus/MapChi
}
