% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_prep.R
\name{census_prep}
\alias{census_prep}
\title{Prepare Data for Geocoding with Addresses Using the Census Bureau API}
\usage{
census_prep(.data, address, id = NA, city = NA, state = NA,
  zip = NA)
}
\arguments{
\item{.data}{Name of the data.frame containing address component variables}

\item{address}{Name of column containing house number and street name}

\item{id}{Name of column containing unique identifier for every observation}

\item{city}{Name of column containing city}

\item{state}{Name of column containing state}

\item{zip}{Name of column containing zipcode}
}
\description{
Prepare Data for Geocoding with Addresses Using the Census Bureau API
}
