% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{cxy_benchmarks}
\alias{cxy_benchmarks}
\title{Get Current Valid Benchmarks}
\usage{
cxy_benchmarks()
}
\value{
A data.frame containing valid Census Benchmarks
}
\description{
Get Current Valid Benchmarks
}
\examples{
cxy_benchmarks()

}
