% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_geocoder.R
\name{cxy_geocode}
\alias{cxy_geocode}
\title{Geocode Addresses Using the Census Bureau Geocoder}
\usage{
cxy_geocode(.data, id = NA, address, city = NA, state = NA,
  zip = NA, timeout = 30)
}
\arguments{
\item{.data}{Dataframe containing address data}

\item{id}{Unique identifier for}

\item{zip}{}
}
\description{
Geocode Addresses Using the Census Bureau Geocoder
}
