% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_join.R
\name{cxy_join}
\alias{cxy_join}
\title{Join the geocoding result with original data}
\usage{

}
\arguments{
\item{.data}{data.frame containing Census API Return (What is returned by `cxy_geocode`)}

\item{original}{data.frame containg the original data to join to}

\item{id}{Name of id variable in original data}

\item{exact}{Logical, if TRUE only joins coordinates from Exact Matches}

\item{simple}{Logical, if TRUE only appends longitude (lon) and latitude (lat) columns}
}
\value{
the original data.frame with full data or just coordinates appended
}
\description{
Join the geocoding result with original data
}
