% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.R
\name{cxy_oneline}
\alias{cxy_oneline}
\title{Single Unstructured Address}
\usage{
cxy_oneline(
  address,
  return = "locations",
  benchmark = "Public_AR_Current",
  vintage = NA
)
}
\arguments{
\item{address}{String containing a single line address}

\item{return}{One of 'locations' or 'geographies' See Details.}

\item{benchmark}{Optional ID or Name of Census Benchmark. See Details.}

\item{vintage}{Optional ID or Name of Census Vintage. See Details.}
}
\value{
A data.frame containing matched address or NULL if not matches
}
\description{
Provides access to the oneline single address geocoding API from the US Census Bureau.
}
