% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.R
\name{cxy_single}
\alias{cxy_single}
\title{Geocode Single Structured Address}
\usage{
cxy_single(
  street,
  city = NA,
  state = NA,
  zip = NA,
  return = "locations",
  benchmark = "Public_AR_Current",
  vintage = NA
)
}
\arguments{
\item{street}{String containing street address}

\item{city}{Optional String containing city}

\item{state}{Optional String containing state}

\item{zip}{Optional String or Integer containing 5-digit Zip Code}

\item{return}{One of 'locations' or 'geographies' See Details.}

\item{benchmark}{Optional ID or Name of Census Benchmark. See Details.}

\item{vintage}{Optional ID or Name of Census Vintage. See Details.}
}
\value{
A data.frame containing matched address or NULL if not matches
}
\description{
Provides access to the structured single address geocoding API from the US Census Bureau.
}
