% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{cxy_vintages}
\alias{cxy_vintages}
\title{Get Current Valid Vintages}
\usage{
cxy_vintages(benchmark)
}
\arguments{
\item{benchmark}{Name or ID of Census Benchmark}
}
\value{
A data.frame containing valid Census Vintages for a given benchmark
}
\description{
Get Current Valid Vintages
}
\examples{
cxy_vintages('Public_AR_Current')

}
