% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_homicides.R
\docType{data}
\name{stl_homicides_small}
\alias{stl_homicides_small}
\title{Homicides in the City of St. Louis July, 2018}
\format{A tibble with 24 rows and 6 variables:
\describe{
  \item{street_address}{number, street and street suffix where homicide occured}
  \item{year}{year homicide occurred}
  \item{date}{data homicide occurred}
  \item{state}{state abbreviation of location, in these data, all "MO"}
  \item{postal_code}{zipcode/postal code of location, in these data all NA}
  \item{city}{city of location, in these data all "St. Louis"}
  }}
\source{
\href{http://www.slmpd.org/Crimereports.shtml}{St. Louis Metropolitan Police Department}
}
\usage{
data(stl_homicides_small)
}
\description{
An example data set containing the addresses for homicides reported by the Saint Louis Metropolitan Police Department
}
\examples{
str(stl_homicides_small)
head(stl_homicides_small)

}
\keyword{datasets}
