% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxy_geocode.R
\name{cxy_geocode}
\alias{cxy_geocode}
\title{Geocode Addresses Using the Census Bureau Geocoder}
\usage{
cxy_geocode(.data, id, address, city, state, zip, timeout = 30, output = "tibble")
}
\arguments{
\item{.data}{dataframe containing address data. Mandatory if using column
names}

\item{id}{Optional; unique identifier for each observation}

\item{address}{Column name containing address}

\item{city}{Optional; column name containing city}

\item{state}{Optional; column name containing state}

\item{zip}{Optional; column name containing 5-digit zip code}

\item{timeout}{Maximum number of minutes for each API call to the geocoder.}

\item{output}{One of either "tibble" or "sf"}
}
\value{
Either a tibble or sf object containing the census geocoder response.
}
\description{
This is the single function of the censusxy package, allowing
    for the easy geocoding of US Addresses using the US Census Bureau
    Geocoder. This function allows for flexible input and virtually
    unlimited batch sizes. See the vignette \code{vignette(censusxy)}
    for more details
}
