% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_geo.R
\name{cxy_geocode}
\alias{cxy_geocode}
\title{Geocode Address Batches Using the Census Bureau API}
\usage{
cxy_geocode(.data, start = 1)
}
\arguments{
\item{.data}{A properly formatted data.frame containing addresses to geocode}

\item{start}{Integer for the first thousand rows to begin geocoding}
}
\description{
It is recommended that `cxy_geocode()` be used in a pipe with `cxy_prep()`. See `vignette("censusxy")` for more details and example code.
When 1,000 or more addresses are specified, the function will parse the data into smaller chunks. The start option allows you to resume geocoding at a specific index.
}
