% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_geocoder.R
\name{cxy_geocode}
\alias{cxy_geocode}
\title{Geocode Addresses Using the Census Bureau Geocoder}
\usage{
cxy_geocode(.data, id = NA, address, city = NA, state = NA,
  zip = NA, timeout = 30, output = "tibble")
}
\arguments{
\item{.data}{dataframe containing address data. Mandatory if using column names}

\item{id}{(Optional) Unique identifier for each observation}

\item{address}{Either a character vector or column name containing address}

\item{city}{(Optional) Either a character vector or column name containing city}

\item{state}{(Optional) Either a character vector or column name containing state}

\item{zip}{(Optional) Either a character/numeric vector or column name containing 5-digit zip code}

\item{timeout}{Maximum number of minutes for each API call to the geocoder.}

\item{output}{Either "tibble" or "sf"}
}
\value{
either a tibble or sf object containing the census geocoder response
}
\description{
This is the single function of the censusxy package, allowing for the easy geocoding of US Addresses using the US Census Bureau Geocoder. This function allows for flexible input and virtually unlimited batch sizes. See the vignette \code{vignette(censusxy)} for more details
}
